requirements.txt
streamlit==1.43.2
pandas==2.2.3
pycaret==3.3.2
python-dotenv==1.0.1
instructor==1.3.7
pydantic==2.10.5
langfuse==2.51.4
openai==1.107.1
